+++ Magic/CHANGELOG.md
@@ -47,6 +47,7 @@
  - Add `earns_multiplier` option to spells, for easily scaling SP earn
  - Add `arenas.enabled false` config optiont to disable Magic Arenas features
  - Automatically generate maybe-unique keys for item attributes, add optional `key` config to override
+ - Add `mana_max_boost` and `mana_regeneration_boost` custom lore options for wands
 
 # 10.11.1